requirements.txt
requests
statistics
python-dotenv
